+++ __tests__/main.ts
@@ -1,4 +1,5 @@
-import { Canvas } from '../packages/g';
+import * as lil from 'lil-gui';
+import { Canvas, CanvasEvent } from '../packages/g';
 import { Renderer as CanvasRenderer } from '../packages/g-canvas';
 import { Renderer as CanvaskitRenderer } from '../packages/g-canvaskit';
 import { Renderer as SVGRenderer } from '../packages/g-svg';
@@ -52,7 +53,11 @@ selectChart.style.margin = '1em';
 renderOptions();
 selectChart.onchange = () => {
   const { value } = selectChart;
-  history.pushState({ value }, '', `?name=${value}`);
+  history.pushState(
+    { value },
+    '',
+    `?name=${value}&renderer=${selectRenderer.value}`,
+  );
   plot();
 };
 document.onkeydown = (event) => {
@@ -86,6 +91,12 @@ const selectRenderer = document.createElement('select');
 selectRenderer.style.margin = '1em';
 selectRenderer.append(...Object.keys(renderers).map(createOption));
 selectRenderer.onchange = () => {
+  const { value } = selectRenderer;
+  history.pushState(
+    { value },
+    '',
+    `?name=${selectChart.value}&renderer=${value}`,
+  );
   plot();
 };
 
@@ -111,7 +122,12 @@ addEventListener('popstate', (event) => {
 
 // @ts-ignore
 const initialValue = new URL(location).searchParams.get('name') as string;
+// @ts-ignore
+const initialRenderer = new URL(location).searchParams.get(
+  'renderer',
+) as string;
 if (tests[initialValue]) selectChart.value = initialValue;
+if (renderers[initialRenderer]) selectRenderer.value = initialRenderer;
 app.append(selectChart);
 app.append(searchInput);
 app.append(selectRenderer);
@@ -167,7 +183,13 @@ function createSpecRender(object) {
         ],
         // Used for WebGPU renderer
         shaderCompilerPath: '/glsl_wgsl_compiler_bg.wasm',
+        // enableDirtyRectangleRendering: false,
+        // enableDirtyRectangleRenderingDebug: true,
       });
+
+      if (generate.initRenderer) {
+        generate.initRenderer(renderer, selectRenderer.value);
+      }
 
       renderer.registerPlugin(
         new DragAndDropPlugin({ dragstartDistanceThreshold: 1 }),
@@ -184,7 +206,36 @@ function createSpecRender(object) {
       // @ts-ignore
       window.__g_instances__ = [canvas];
 
-      await generate({ canvas, renderer, container: $div });
+      // GUI
+      const gui = new lil.GUI({ autoPlace: false });
+      $div.appendChild(gui.domElement);
+
+      await generate({ canvas, renderer, container: $div, gui });
+
+      if (
+        selectRenderer.value === 'canvas' &&
+        renderer.getConfig().enableDirtyRectangleRendering &&
+        renderer.getConfig().enableDirtyRectangleRenderingDebug
+      ) {
+        // display dirty rectangle
+        const $dirtyRectangle = document.createElement('div');
+        $dirtyRectangle.style.cssText = `
+        position: absolute;
+        pointer-events: none;
+        background: rgba(255, 0, 0, 0.5);
+        `;
+        $div.appendChild($dirtyRectangle);
+        canvas.addEventListener(CanvasEvent.DIRTY_RECTANGLE, (e) => {
+          const { dirtyRect } = e.detail;
+          const { x, y, width, height } = dirtyRect;
+          const dpr = window.devicePixelRatio;
+          // convert from canvas coords to viewport
+          $dirtyRectangle.style.left = `${x / dpr}px`;
+          $dirtyRectangle.style.top = `${y / dpr}px`;
+          $dirtyRectangle.style.width = `${width / dpr}px`;
+          $dirtyRectangle.style.height = `${height / dpr}px`;
+        });
+      }
 
       container.append($div);
     };