+++ packages/g-dom-mutation-observer-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @antv/g-dom-mutation-observer-api
 
+## 1.2.22
+
+### Patch Changes
+
+-   Updated dependencies [11d23f39]
+    -   @antv/g-lite@1.2.22
+
 ## 1.2.21
 
 ### Patch Changes