+++ packages/g-plugin-canvas-picker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @antv/g-plugin-canvas-picker
 
+## 1.10.24
+
+### Patch Changes
+
+-   Updated dependencies [11d23f39]
+    -   @antv/g-lite@1.2.22
+    -   @antv/g-plugin-canvas-path-generator@1.3.22
+    -   @antv/g-plugin-canvas-renderer@1.9.24
+
 ## 1.10.23
 
 ### Patch Changes