+++ packages/g-plugin-canvaskit-renderer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/g-plugin-canvaskit-renderer
 
+## 1.3.23
+
+### Patch Changes
+
+-   Updated dependencies [11d23f39]
+    -   @antv/g-lite@1.2.22
+    -   @antv/g-plugin-image-loader@1.3.22
+
 ## 1.3.22
 
 ### Patch Changes