+++ packages/g-plugin-zdog-canvas-renderer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/g-plugin-zdog-canvas-renderer
 
+## 1.2.27
+
+### Patch Changes
+
+-   Updated dependencies [11d23f39]
+    -   @antv/g-lite@1.2.22
+    -   @antv/g-canvas@1.11.27
+
 ## 1.2.26
 
 ### Patch Changes